requirements.txt
requests
pandas
matplotlib
python-telegram-bot
apscheduler
